+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/composable/actions/ComposableEmulate.kt
@@ -15,7 +15,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.flipperdevices.bridge.dao.api.model.FlipperKey
 import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
@@ -24,10 +23,11 @@ import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.keyscreen.impl.R
 import com.flipperdevices.keyscreen.impl.model.EmulateButtonState
 import com.flipperdevices.keyscreen.impl.viewmodel.EmulateViewModel
+import tangle.viewmodel.compose.tangleViewModel
 
 @Composable
 fun ComposableEmulate(modifier: Modifier = Modifier, flipperKey: FlipperKey) {
-    val emulateViewModel = viewModel<EmulateViewModel>()
+    val emulateViewModel = tangleViewModel<EmulateViewModel>()
     val emulateButtonState by emulateViewModel.getEmulateButtonStateFlow().collectAsState()
     if (!flipperKey.synchronized) {
         ComposableActionDisable(