+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/di/KeyScreenComponent.kt
@@ -2,7 +2,6 @@ package com.flipperdevices.keyscreen.impl.di
 
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.keyscreen.impl.fragments.KeyScreenFragment
-import com.flipperdevices.keyscreen.impl.viewmodel.EmulateViewModel
 import com.flipperdevices.keyscreen.impl.viewmodel.FlipperDeviceViewModel
 import com.flipperdevices.keyscreen.impl.viewmodel.KeyScreenViewModel
 import com.squareup.anvil.annotations.ContributesTo
@@ -11,6 +10,5 @@ import com.squareup.anvil.annotations.ContributesTo
 interface KeyScreenComponent {
     fun inject(fragment: KeyScreenFragment)
     fun inject(viewModel: KeyScreenViewModel)
-    fun inject(viewModel: EmulateViewModel)
     fun inject(viewModel: FlipperDeviceViewModel)
 }